requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
pyyaml>=6.0
aiofiles>=23.0
emoji>=2.0
